+++ app/views/avo/partials/_table_header.html.erb
@@ -20,7 +20,7 @@
       <% end %>
     </th>
   <% end %>
-  <% if @resource.resource_controls_render_on_the_left? %>
+  <% if @resource.render_row_controls_on_the_left? %>
     <th class="max-w-24" data-control="resource-controls-th">
       <!-- Item controls cell -->
     </th>
@@ -105,7 +105,7 @@
       <% end %>
     <% end %>
   <% end %>
-  <% if @resource.resource_controls_render_on_the_right? %>
+  <% if @resource.render_row_controls_on_the_right? %>
     <th class="w-px max-w-24" data-control="resource-controls-th">
       <!-- Item controls cell -->
     </th>
