+++ spec/dummy/.env.sample
@@ -9,5 +9,3 @@ AWS_ACCESS_KEY_ID=
 AWS_REGION=
 AWS_S3_BUCKET=
 AWS_SECRET_ACCESS_KEY=
-
-# AVO_RESOURCE_CONTROLS_PLACEMENT=right
