+++ spec/dummy/app/avo/resources/user.rb
@@ -45,10 +45,10 @@ class Avo::Resources::User < Avo::BaseResource
     end
   }
 
-  # self.row_controls_config = {
-  #   float: true,
-  #   show_on_hover: true
-  # }
+  self.row_controls_config = {
+    float: true,
+    show_on_hover: true
+  }
 
   def fields
     test_field("Heading")
